fix: deps
